requirements.txt
flask==3.0.0
requests==2.31.0
yt-dlp==2024.3.10
moviepy==1.0.3
pillow==10.2.0
instagrapi==2.1.0
opencv-python==4.9.0.80
python-dotenv==1.0.1
imageio-ffmpeg==0.4.9
numpy==1.26.4
